rename the library from scpi to scpilib as the name is already in use in pypi.
